requirements.txt
requests=2.28.*
python_dotenv=0.21.*